+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - 🛠️ Fixed page titles not working properly.
 - ⚙️ Removed redundant HTML templates, now using index.html as template for all pages.
 - ⚙️ CreateElement utility function now handles CSS variables, they can be passed directly as a prop instead of using element.style.setProperty() on returned object.
+- ⚙️ Button UI component now exports an object instead of the element directly.
 
 ### **2.0.1 (8 May, 2025)**